Add files via upload
Given the interests of a student this program will help suggest suitable streams and subjects for students going to class 11.
Just enter the name of the student and the stream you want to check in the format study(student_name, stream). if you want to check if that stream is suitable.
Or enter study(student_name, X). this will return the streams which are suitable from the pool, type a semicolon to see if more than 1 stream is suitable.
The program will return Yes or No or the name of the stream based on the subjects of interest of the student.